# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SWP Summit Conference Registration API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pricing/calculate:
    post:
      operationId: calculatePricing
      tags: [pricing]
      summary: Calculate pricing for a given pass type and quantity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingRequest"
      responses:
        "200":
          description: Pricing breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingBreakdown"

  # ==================== PROMO CODES ====================
  /promo-codes/validate:
    post:
      operationId: validatePromoCode
      tags: [promo-codes]
      summary: Validate a promo code for a given booking context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidatePromoCodeBody"
      responses:
        "200":
          description: Promo code is valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCodeValidationResult"
        "400":
          description: Invalid or expired promo code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/sponsors:
    get:
      operationId: listSponsors
      tags: [sponsors-admin]
      summary: List sponsor records with progress and allocation usage
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/SponsorStatus" } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Sponsor list
          content:
            application/json:
              schema:
                type: object
                properties:
                  sponsors: { type: array, items: { $ref: "#/components/schemas/SponsorSummary" } }
                required: [sponsors]
    post:
      operationId: createSponsor
      tags: [sponsors-admin]
      summary: Create a draft sponsor record
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SponsorUpsert" }
      responses:
        "201":
          description: Draft sponsor created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SponsorWorkspaceAdmin" }
        "409":
          {
            description: Promo code collision,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } },
          }
  /admin/sponsors/{sponsorId}:
    parameters:
      - { $ref: "#/components/parameters/SponsorId" }
    get:
      operationId: getAdminSponsor
      tags: [sponsors-admin]
      summary: Get a complete sponsor administration record
      responses:
        "200":
          {
            description: Sponsor record,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/SponsorWorkspaceAdmin" } },
              },
          }
    patch:
      operationId: updateAdminSponsor
      tags: [sponsors-admin]
      summary: Update sponsor details, allocations, contacts, deadlines or status
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SponsorUpsert" }
      responses:
        "200":
          {
            description: Updated sponsor,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/SponsorWorkspaceAdmin" } },
              },
          }
        "409":
          {
            description: Allocation below active use or promo collision,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } },
          }
...
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
  parameters:
    SponsorId:
      name: sponsorId
      in: path
      required: true
      schema: { type: integer }
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SponsorStatus:
      type: string
      enum: [draft, confirmed, paused, completed, cancelled]
    SponsorAssetCategory:
      type: string
      enum: [logo, headshot, slides, session_material, logistics, other]
    SponsorContact:
      type: object
      properties:
        id: { type: integer }
        role: { type: string, enum: [primary, onsite, marketing, other] }
        firstName: { type: string }
        lastName: { type: string }
        jobTitle: { type: string, nullable: true }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        isPrimary: { type: boolean }
      required: [role, firstName, lastName, email, isPrimary]
    SponsorTask:
      type: object
      properties:
        id: { type: integer }
        taskKey: { type: string }
        label: { type: string }
        required: { type: boolean }
        dueAt: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [todo, submitted, completed, overdue, not_required] }
        completedAt: { type: string, format: date-time, nullable: true }
      required: [id, taskKey, label, required, status]
    SponsorUsagePerson:
      type: object
      description: Privacy-redacted usage; contact and financial fields are intentionally absent.
      properties:
        bookingId: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        company: { type: string }
        jobTitle: { type: string }
        registeredAt: { type: string, format: date-time }
      required: [bookingId, firstName, lastName, company, jobTitle, registeredAt]
    SponsorCode:
      type: object
      properties:
        kind: { type: string, enum: [vip, public] }
        code: { type: string }
        active: { type: boolean }
        workforceUrl: { type: string, format: uri }
        allocation: { type: integer, nullable: true }
        used: { type: integer }
        remaining: { type: integer, nullable: true }
        maxPerBooking: { type: integer, nullable: true }
        discountPercent: { type: number, nullable: true }
        redemptions: { type: array, items: { $ref: "#/components/schemas/SponsorUsagePerson" } }
      required: [kind, code, active, workforceUrl, used, redemptions]
    SponsorStaff:
      type: object
      properties:
        bookingId: { type: integer }
        attendeeId: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        jobTitle: { type: string }
        company: { type: string }
        workEmail: { type: string, format: email }
        phone: { type: string, nullable: true }
        dietaryAccessibility: { type: string, nullable: true }
        communitySocialAttending: { type: boolean, nullable: true }
        communitySocialDietary: { type: string, nullable: true }
        marketingConsent: { type: boolean }
        status: { type: string }
        registeredAt: { type: string, format: date-time }
      required:
        [
          bookingId,
          attendeeId,
          firstName,
          lastName,
          jobTitle,
          company,
          workEmail,
          marketingConsent,
          status,
          registeredAt,
        ]
...
    SponsorPresenter:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        jobTitle: { type: string }
        company: { type: string }
        biography: { type: string, nullable: true }
        displayOrder: { type: integer }
      required: [name, jobTitle, company]
    SponsorSession:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [quickfire, keynote, other] }
        entitlementLabel: { type: string }
        title: { type: string, nullable: true }
        description: { type: string, nullable: true }
        takeaways: { type: array, maxItems: 3, items: { type: string } }
        status: { type: string, enum: [draft, submitted, changes_requested, approved, exported] }
        feedback: { type: string, nullable: true }
        headshotRequired: { type: boolean }
        takeawaysRequired: { type: boolean }
        slidesRequired: { type: boolean }
        currentRevision: { type: integer }
        exportedRevision: { type: integer, nullable: true }
        exportOutdated: { type: boolean }
        presenters: { type: array, items: { $ref: "#/components/schemas/SponsorPresenter" } }
        revisions:
          type: array
          items:
            type: object
            properties:
              revision: { type: integer }
              actor: { type: string }
              createdAt: { type: string, format: date-time }
            required: [revision, actor, createdAt]
      required:
        [
          id,
          type,
          entitlementLabel,
          takeaways,
          status,
          headshotRequired,
          takeawaysRequired,
          slidesRequired,
          currentRevision,
          exportOutdated,
          presenters,
          revisions,
        ]
...
    SponsorAsset:
      type: object
      properties:
        id: { type: string }
        sponsorId: { type: integer }
        sponsorCompany: { type: string, nullable: true }
        sessionId: { type: integer, nullable: true }
        presenterId: { type: integer, nullable: true }
        category: { $ref: "#/components/schemas/SponsorAssetCategory" }
        originalName: { type: string }
        mimeType: { type: string }
        byteSize: { type: integer }
        checksumSha256: { type: string }
        version: { type: integer }
        status: { type: string, enum: [active, archived, missing] }
        uploaderType: { type: string }
        uploaderLabel: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        previewAvailable: { type: boolean }
      required:
        [
          id,
          sponsorId,
          category,
          originalName,
          mimeType,
          byteSize,
          checksumSha256,
          version,
          status,
          uploaderType,
          createdAt,
          previewAvailable,
        ]
...
    SponsorDocument:
      type: object
      properties:
        id: { type: integer }
        assetId: { type: string }
        title: { type: string }
        required: { type: boolean }
        acknowledgementVersion: { type: integer }
        acknowledged: { type: boolean }
        acknowledgedBy: { type: string, nullable: true }
        acknowledgedAt: { type: string, format: date-time, nullable: true }
      required: [id, assetId, title, required, acknowledgementVersion, acknowledged]
    SponsorSummary:
      type: object
      properties:
        id: { type: integer }
        company: { type: string }
        packageLabel: { type: string }
        status: { $ref: "#/components/schemas/SponsorStatus" }
        confirmationDate: { type: string, format: date, nullable: true }
        vipAllocation: { type: integer }
        vipUsed: { type: integer }
        staffAllocation: { type: integer }
        staffUsed: { type: integer }
        progressCompleted: { type: integer }
        progressTotal: { type: integer }
        needsAttention: { type: integer }
        updatedAt: { type: string, format: date-time }
      required:
        [
          id,
          company,
          packageLabel,
          status,
          vipAllocation,
          vipUsed,
          staffAllocation,
          staffUsed,
          progressCompleted,
          progressTotal,
          needsAttention,
          updatedAt,
        ]
    SponsorUpsert:
      type: object
      properties:
        company: { type: string, minLength: 1, maxLength: 200 }
        packageLabel: { type: string, minLength: 1, maxLength: 200 }
        status: { $ref: "#/components/schemas/SponsorStatus" }
        confirmationDate: { type: string, format: date, nullable: true }
        notes: { type: string, nullable: true, maxLength: 10000 }
        vipAllocation: { type: integer, minimum: 0 }
        vipMaxPerBooking: { type: integer, minimum: 1 }
        staffAllocation: { type: integer, minimum: 0 }
        vipCode: { type: string, pattern: "^[A-Z0-9]+$" }
        publicCode: { type: string, pattern: "^[A-Z0-9]+$" }
        contacts: { type: array, items: { $ref: "#/components/schemas/SponsorContact" } }
        tasks: { type: array, items: { $ref: "#/components/schemas/SponsorTask" } }
        sessions:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [quickfire, keynote, other] }
              entitlementLabel: { type: string }
              headshotRequired: { type: boolean }
              takeawaysRequired: { type: boolean }
              slidesRequired: { type: boolean }
            required: [type, entitlementLabel]
      required: [company, packageLabel, vipAllocation, vipMaxPerBooking, staffAllocation]
    SponsorWorkspaceAdmin:
      allOf:
        - { $ref: "#/components/schemas/SponsorSummary" }
        - type: object
          properties:
            notes: { type: string, nullable: true }
            accessUrl: { type: string, format: uri, nullable: true }
            welcomeEmailSentAt: { type: string, format: date-time, nullable: true }
            contacts: { type: array, items: { $ref: "#/components/schemas/SponsorContact" } }
            codes: { type: array, items: { $ref: "#/components/schemas/SponsorCode" } }
            staff: { type: array, items: { $ref: "#/components/schemas/SponsorStaff" } }
            tasks: { type: array, items: { $ref: "#/components/schemas/SponsorTask" } }
            sessions: { type: array, items: { $ref: "#/components/schemas/SponsorSession" } }
            assets: { type: array, items: { $ref: "#/components/schemas/SponsorAsset" } }
            documents: { type: array, items: { $ref: "#/components/schemas/SponsorDocument" } }
            activity: { type: array, items: { type: object, additionalProperties: true } }
          required: [contacts, codes, staff, tasks, sessions, assets, documents, activity]
...
    PricingRequest:
      type: object
      properties:
        passType:
          type: string
          enum: [single, business]
        quantity:
          type: integer
        promoCode:
          type: ["string", "null"]
      required:
        - passType
        - quantity
    PricingBreakdown:
      type: object
      properties:
        passType:
          type: string
        quantity:
          type: integer
        pricePerHead:
          type: number
        baseSubtotal:
          type: number
        groupDiscountPercent:
          type: number
        groupDiscountAmount:
          type: number
        promoDiscountAmount:
          type: number
        subtotalAfterDiscounts:
          type: number
        vatRate:
          type: number
        vatAmount:
          type: number
        total:
          type: number
        originalPrice:
          type: number
        savedAmount:
          type: number
        promoDiscountType:
          type: ["string", "null"]
          enum: [percentage, fixed, per_ticket, complimentary, null]
        promoRemainingSeats:
          type: ["integer", "null"]
      required:
        - passType
        - quantity
        - pricePerHead
        - baseSubtotal
        - groupDiscountPercent
        - groupDiscountAmount
        - promoDiscountAmount
        - subtotalAfterDiscounts
        - vatRate
        - vatAmount
        - total
        - originalPrice
        - savedAmount

    # ==================== PROMO CODE SCHEMAS ====================
...
    ValidatePromoCodeBody:
      type: object
      properties:
        code:
          type: string
        passType:
          type: string
          enum: [single, business]
        quantity:
          type: integer
        leadEmail:
          type: ["string", "null"]
      required:
        - code
        - passType
        - quantity
    PromoCodeValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed, per_ticket, complimentary]
        discountValue:
          type: number
        discountAmount:
          type: number
        message:
          type: ["string", "null"]
        remainingSeats:
          type: ["integer", "null"]
          description: For complimentary codes only — the number of free tickets still available on this code (null if the code is uncapped or not complimentary).
      required:
        - valid
        - code
        - discountType
        - discountValue
        - discountAmount

    # ==================== DISCOUNT TIER SCHEMAS ====================
...